tests(cli) New CLI tests (start/stop)

+++ spec/helpers.lua
@@ -1,15 +1,24 @@
 local BIN_PATH = "bin/kong"
 local TEST_CONF_PATH = "spec/kong_tests.conf"
-local TEST_PREFIX_PATH = "servroot_tests"
 
 local conf_loader = require "kong.conf_loader"
+local DAOFactory = require "kong.dao.factory"
 local pl_utils = require "pl.utils"
 local pl_path = require "pl.path"
 local pl_dir = require "pl.dir"
 
+---------------
+-- Conf and DAO
+---------------
 local conf = assert(conf_loader(TEST_CONF_PATH))
+local dao = DAOFactory(conf)
 
+-- make sure migrations are up-to-date
+assert(dao:run_migrations())
+
+--------------------
 -- Custom assertions
+--------------------
 local say = require "say"
 local luassert = require "luassert.assert"
 
@@ -44,6 +53,9 @@ Body:
 luassert:register("assertion", "res_status", res_status,
                   "assertion.res_status.negative")
 
+----------------
+-- Shell helpers
+----------------
 local function exec(...)
   local ok, _, _, stderr = pl_utils.executeex(...)
   return ok, stderr
@@ -51,30 +63,37 @@ end
 
 local function kong_exec(args, prefix)
   args = args or ""
-  prefix = prefix or TEST_PREFIX_PATH
+  prefix = prefix or conf.prefix
+
   return exec(BIN_PATH.." "..args.." --prefix "..prefix)
 end
 
+----------
+-- Exposed
+----------
 return {
   -- Penlight
   dir = pl_dir,
   path = pl_path,
   execute = pl_utils.executeex,
+
   -- Kong testing properties
   bin_path = BIN_PATH,
   test_conf = conf,
-  test_prefix = TEST_PREFIX_PATH,
   test_conf_path = TEST_CONF_PATH,
+
   -- Kong testing helpers
   kong_exec = kong_exec,
   prepare_prefix = function(prefix)
-    prefix = prefix or TEST_PREFIX_PATH
+    prefix = prefix or conf.prefix
     pl_dir.makepath(prefix)
     kong_exec("stop", prefix)
   end,
   clean_prefix = function(prefix)
-    prefix = prefix or TEST_PREFIX_PATH
-    pl_dir.rmtree(prefix)
+    prefix = prefix or conf.prefix
+    if pl_path.exists(prefix) then
+      pl_dir.rmtree(prefix)
+    end
   end,
   start_kong = function(prefix)
     return kong_exec("start --conf "..TEST_CONF_PATH, prefix)
